lambda-egyptian-api/requirements.txt
google-generativeai==0.8.5
Pillow==10.4.0
pydantic==2.10.4
